pyproject.toml
[tool.poetry]
name = "networkdevicesmetrics"
version = "0.1.0"
description = "Сбор и проверка метрик сетевых устройств"
authors = ["Dmitry <newrelax@rambler.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "5.2.*"
djangorestframework = "3.16.*"
django-filter = "25.1.*"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
[dependency-groups]
dev = [
    "ruff (>=0.15.11,<0.16.0)"
]
